improved auth, added progress filtering

+++ app/Http/Controllers/TodoController.php
@@ -38,8 +38,14 @@ class TodoController extends Controller
         $data = $request-> validate ([
             'title' => 'required |string',
             'completed' => 'required|boolean',
+            'inProgress' => 'required|boolean'
         ]);
-        $todo = Todo:: create($data);
+        $todo = Todo:: create([
+            'user_id' => auth()->user()->id,
+            'title'=> $request->title,
+            'completed' => $request->completed,
+            'inProgress' => $request->inProgress,
+        ]);
         return response($todo);
     }
 
@@ -74,9 +80,13 @@ class TodoController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if($todo->user_id !== auth()->user()->id){
+            return response()->json('Unauthorized', 401);
+        }
         $data = $request-> validate ([
             'title' => 'required |string',
             'completed' => 'required|boolean',
+            'inProgress' => 'required|boolean'
         ]);
         $todo->update($data);
         return response($todo, 200);
@@ -87,7 +97,7 @@ class TodoController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
 
         return response("Updated", 200);
     }
@@ -100,6 +110,9 @@ class TodoController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if($todo->user_id !== auth()->user()->id){
+            return response()->json('Unauthorized', 401);
+        }
         $todo->delete();
         return response("Deleted selected task", 200);
     }
@@ -108,11 +121,21 @@ class TodoController extends Controller
 
     public function destroyCompleted(Request $request)
     {
+        $todosToDelete = $request->todos;
+        $userTodoIds = auth()->user()->todos->map(function ($todo){
+            return $todo->id;
+        });
+        $valid = collect($todosToDelete)->every(function($value, $key) use ($userTodoIds){
+            return $userTodoIds->contains($value);
+        });
+        if(!$valid){
+            return response()->json('Unauthorized', 401);
+        }
         $request->validate([
             'todos' => 'required|array',
         ]);
         Todo::destroy($request->todos);
-        return response("Deleted all completed tasks");
+        return response()->json("Deleted all completed tasks", 200);
     }
 
 }
